Update model_types to model_variable.h
